scripts/requirements.txt
anthropic>=0.40.0
PyGithub>=2.0.0
